+++ src/mite_extras/processing/validation_manager.py
@@ -49,8 +49,9 @@ logger = logging.getLogger("mite_extras")
 class IdValidator(BaseModel):
     """Handles validation and cross-referencing of database identifiers."""
 
+    @staticmethod
     def cleanup_ids(
-        self, genpept: str | None = None, uniprot: str | None = None
+        genpept: str | None = None, uniprot: str | None = None
     ) -> dict[str, str]:
         """Cleans up IDs using the UniProt SPARQL endpoint.
 
@@ -146,6 +147,9 @@ class IdValidator(BaseModel):
 
         Args:
             qid: a valid Wikidata QID
+
+        Raises:
+            ValueError: Wikidata QID invalid or could not be retrieved
         """
 
         def build_query(qid: str) -> str:
@@ -172,7 +176,9 @@ class IdValidator(BaseModel):
             return data.get("boolean")
 
         if not fetch_result(query=build_query(qid=qid)):
-            logger.warning(f"Wikidata QID '{qid}' does not exist or has no statements.")
+            raise ValueError(
+                f"Wikidata QID '{qid}' does not exist or has no statements."
+            )
 
 
 class MoleculeValidator(BaseModel):