+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.4] 25-10-2025
+
+### Changed
+
+- Missing/wrong Wikidata QID does not raise error, only logs warning (matches genpept/uniprot behavior)
+
 ## [1.6.3] 30-09-2025
 
 ### Changed